+++ src/components/PasswordInput/PasswordInput.styles.ts
@@ -1,0 +1,13 @@
+import styled from '@emotion/native';
+
+const EyeIcon = styled.View`
+  position: absolute;
+  top: 17px;
+  right: 17px;
+`;
+
+const S = {
+  EyeIcon,
+};
+
+export default S;
